moved calojet timing filter and producer to plugins
