+++ HLTrigger/JetMET/plugins/HLTCaloJetTimingFilter.cc
@@ -1,12 +1,60 @@
 /** \class HLTCaloJetTimingFilter
  *
- * See header file for documentation
- *
+ *  \brief  This makes selections on the timing and associated ecal cells 
+ *  produced by HLTCaloJetTimingProducer
  *  \author Matthew Citron
+ *
  *
  */
 
-#include "HLTrigger/JetMET/interface/HLTCaloJetTimingFilter.h"
+// system include files
+#include <memory>
+
+// user include files
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "HLTrigger/HLTcore/interface/HLTFilter.h"
+
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/StreamID.h"
+
+#include "DataFormats/HLTReco/interface/TriggerFilterObjectWithRefs.h"
+
+namespace edm {
+  class ConfigurationDescriptions;
+}
+
+//
+// class declaration
+//
+class HLTCaloJetTimingFilter : public HLTFilter {
+public:
+  explicit HLTCaloJetTimingFilter(const edm::ParameterSet& iConfig);
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+  bool hltFilter(edm::Event&,
+                 const edm::EventSetup&,
+                 trigger::TriggerFilterObjectWithRefs& filterproduct) const override;
+
+private:
+  //Input collections
+  edm::InputTag jetLabel_;
+  edm::InputTag jetTimeLabel_;
+  edm::InputTag jetCellsForTimingLabel_;
+  edm::InputTag jetEcalEtForTimingLabel_;
+  //Thresholds for selection
+  unsigned int minJets_;
+  double jetTimeThresh_;
+  double jetEcalEtForTimingThresh_;
+  unsigned int jetCellsForTimingThresh_;
+  double minPt_;
+
+  edm::EDGetTokenT<reco::CaloJetCollection> jetInputToken;
+  edm::EDGetTokenT<edm::ValueMap<float>> jetTimesInputToken;
+  edm::EDGetTokenT<edm::ValueMap<unsigned int>> jetCellsForTimingInputToken;
+  edm::EDGetTokenT<edm::ValueMap<float>> jetEcalEtForTimingInputToken;
+};
 
 //Constructor
 HLTCaloJetTimingFilter::HLTCaloJetTimingFilter(const edm::ParameterSet& iConfig) : HLTFilter(iConfig) {
@@ -69,3 +117,6 @@ void HLTCaloJetTimingFilter::fillDescriptions(edm::ConfigurationDescriptions& de
   desc.add<double>("minJetPt", 40.);
   descriptions.add("caloJetTimingFilter", desc);
 }
+
+// declare this class as a framework plugin
+DEFINE_FWK_MODULE(HLTCaloJetTimingFilter);
