+++ HLTrigger/JetMET/plugins/HLTCaloJetTimingProducer.cc
@@ -1,12 +1,57 @@
 /** \class HLTCaloJetTimingProducer
  *
- * See header file for documentation
- *
+ *  \brief  This produces timing and associated ecal cell information for calo jets 
  *  \author Matthew Citron
+ *
  *
  */
 
-#include "HLTrigger/JetMET/interface/HLTCaloJetTimingProducer.h"
+// system include files
+#include <memory>
+
+// user include files
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/StreamID.h"
+#include "DataFormats/Common/interface/ValueMap.h"
+
+#include "DataFormats/JetReco/interface/CaloJetCollection.h"
+
+#include "Geometry/CaloGeometry/interface/CaloGeometry.h"
+#include "Geometry/Records/interface/CaloGeometryRecord.h"
+#include "DataFormats/EcalRecHit/interface/EcalRecHitCollections.h"
+#include "HLTrigger/HLTcore/interface/defaultModuleLabel.h"
+#include "TLorentzVector.h"
+#include "DataFormats/Math/interface/deltaR.h"
+
+//
+// class declaration
+//
+class HLTCaloJetTimingProducer : public edm::stream::EDProducer<> {
+public:
+  explicit HLTCaloJetTimingProducer(const edm::ParameterSet&);
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  void produce(edm::Event&, const edm::EventSetup&) override;
+
+  // ----------member data ---------------------------
+  // Input collections
+  edm::InputTag jetLabel_;
+  edm::InputTag ecalEBLabel_;
+  edm::InputTag ecalEELabel_;
+  // Include endcap jets or only barrel
+  bool barrelOnly_;
+
+  edm::EDGetTokenT<reco::CaloJetCollection> jetInputToken;
+  edm::EDGetTokenT<edm::SortedCollection<EcalRecHit, edm::StrictWeakOrdering<EcalRecHit>>> ecalRecHitsEBToken;
+  edm::EDGetTokenT<edm::SortedCollection<EcalRecHit, edm::StrictWeakOrdering<EcalRecHit>>> ecalRecHitsEEToken;
+};
 
 //Constructor
 HLTCaloJetTimingProducer::HLTCaloJetTimingProducer(const edm::ParameterSet& iConfig) {
@@ -122,3 +167,6 @@ void HLTCaloJetTimingProducer::fillDescriptions(edm::ConfigurationDescriptions& 
   desc.add<edm::InputTag>("eeRecHitsColl", edm::InputTag("hltEcalRecHit", "EcalRecHitsEE"));
   descriptions.add("caloJetTimingProducer", desc);
 }
+
+// declare this class as a framework plugin
+DEFINE_FWK_MODULE(HLTCaloJetTimingProducer);
